refactor(proactive): share tool roster wiring
