+++ packages/lib/proactive/src/create-proactive-tools.test.ts
@@ -1,21 +1,12 @@
 import { describe, expect, test } from "bun:test";
-import { createProactiveTools } from "./create-proactive-tools.js";
+import { createProactiveTools, PROACTIVE_TOOL_NAMES } from "./create-proactive-tools.js";
 import { createSchedulerStub } from "./test-helpers.js";
 
 describe("createProactiveTools", () => {
   test("returns sleep, cron, and monitor tools in stable order", () => {
     const stub = createSchedulerStub();
     const tools = createProactiveTools({ scheduler: stub.component });
-    expect(tools.map((t) => t.descriptor.name)).toEqual([
-      "sleep",
-      "cancel_sleep",
-      "schedule_cron",
-      "cancel_schedule",
-      "create_monitor",
-      "list_monitors",
-      "update_monitor",
-      "cancel_monitor",
-    ]);
+    expect(tools.map((t) => t.descriptor.name)).toEqual(PROACTIVE_TOOL_NAMES);
   });
 
   test("all tools share the primordial origin", () => {
