+++ packages/lib/proactive/src/provider.test.ts
@@ -1,6 +1,7 @@
 import { describe, expect, test } from "bun:test";
 import type { Agent, SchedulerComponent, SubsystemToken } from "@koi/core";
 import { COMPONENT_PRIORITY, SCHEDULER, toolToken } from "@koi/core";
+import { PROACTIVE_TOOL_NAMES } from "./create-proactive-tools.js";
 import { createProactiveToolsProvider } from "./provider.js";
 import { createSchedulerStub } from "./test-helpers.js";
 
@@ -63,19 +64,9 @@ describe("createProactiveToolsProvider", () => {
 
     const result = await provider.attach(agent);
     const components = "components" in result ? result.components : result;
-    const toolNames = [
-      "sleep",
-      "cancel_sleep",
-      "schedule_cron",
-      "cancel_schedule",
-      "create_monitor",
-      "list_monitors",
-      "update_monitor",
-      "cancel_monitor",
-    ] as const;
-    for (const n of toolNames) {
-      expect(components.has(toolToken(n) as string)).toBe(true);
-    }
+    const expectedKeys = PROACTIVE_TOOL_NAMES.map((name) => toolToken(name) as string);
+    expect(components.size).toBe(expectedKeys.length);
+    expect([...components.keys()]).toEqual(expectedKeys);
   });
 
   test("attach surfaces a skipped entry when the agent has no SCHEDULER component", async () => {
